Add files via upload
Simulation of Cord and Li attractors

Here I provide Matlab and Julia scripts to simulate the Cord and Li attractors. The numerical integration is performed by fixed-step 4th-order Runge-Kutta algorithm which will not get stuck in the case of unstable systems and easily accommodates input signals (for control and synchronization).