Add files via upload
Baker's Map

Baker's map is a conceptual model that emulates the "stretching and folding"; actions of a baker on a piece of dough. These two elements are responsible for some of the key features of chaotic dynamics, such as extreme sensitivity to initial conditions and global boundedness of the dynamics.