Add files via upload
Network of Kuramoto oscillators

These scripts (three Matlab scripts are compressed in the zip file and a stand alone Julia file) simulate a network of Kuramoto oscillators. The user can grow the network size, change the coupling among oscillators to see how that affects synchronization. The scripts implement a bidirectionally connected chain of oscillators, but the user can change the topology defining the corresponding adjacency matrix.