Add files via upload
Recurrence plots: examples

The zip file attached is a self-contained set of scripts that will produce Recurrence Plots for a periodic signal, a random signal and a chaotic signal. The scripts RPsignals.m and RPCord.m can be easily adapted to produce Recurrence Plots for other signals and systems. Please notice that some scripts are due to: 
%Author: Hui Yang 
%Affiliation: The Pennsylvania State University 310 Leohard Building, University Park, PA 
%Email: yanghui@gmail.com