Add files via upload
Runge Kutta simulation in Python

In the zipped file you will find three stand-alone notebook files for numerical simulation of three nonlinear dynamical systems: classical Rössler, modified Rössler and Sprott H. All of them are autonomous and could be simulated directly using off-the-shelf functions. Instead, I coded the 4th order Runge Kutta integration scheme for each. This will help when simulating these systems with exogenous variables (control and syncronization problems). In that case, it is convenient to be able to send the control actions as input parameters. The codes are basically the same and can easily be adapted in order to simulate any other dynamical system.