Add files via upload
Symbolic sequence and entropy for Rössler spiral attractor

The script RosSymbols.m (inside the zip file) simulates and produces symbolic dynamics (sequences of zeros and ones) for the Rössler spiral attractor. The corresponding cycles are shown on the attractor for an interpretation of what a 0 (or 1) cycle means. This is used in the sequel to estimate the Shannon entropy using 4-bit words.